Rename Prim's Algorithm to Prims_Algorithm.py
Changed the file name to something more usable. The way it was previously could have caused the program to fail to run for certain people. I also changed the ending to .py to signify that it is a python file. This helps in your development so you will have syntax highlighting.